requirements.txt
discord.py>=2.3.2
sentence-transformers>=2.7.0
anthropic>=0.40.0
python-dotenv>=1.0.0
numpy>=1.24.0
